# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This Api is used for creating and managing giftcards",
        "title": "Gift Card API",
        "termsOfService": "NONE",
        "contact": {
            "name": "info@jabama.com",
            "url": "jabama.com",
            "email": "info@jabama.com"
        },
        "license": {
            "name": "NONE",
            "url": "None"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/campaign/page/{size}/{number}": {
            "get": {
                "description": "get list of campaigns in paging object",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Campaign paging",
                "operationId": "find-page",
                "parameters": [
                    {
                        "type": "number",
                        "description": "page size",
                        "name": "size",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "page number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "search by title",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CampaignPageDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
...
        "/v1/gift-card": {
            "put": {
                "description": "updates a gift card. just the expire date and the amount can be updated",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gift Card"
                ],
                "summary": "updates a gift card",
                "operationId": "update",
                "parameters": [
                    {
                        "description": "Update Gift Card dto",
                        "name": "giftCardDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateGiftCardDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GiftCardDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            },
            "post": {
                "description": "store a new gift card and generates the keys",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gift Card"
                ],
                "summary": "store a gift card",
                "operationId": "store",
                "parameters": [
                    {
                        "description": "Create Gift Card dto",
                        "name": "giftCardDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateGiftCardDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GiftCardDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
        "/v1/gift-card/approve-gift-card/{uun}/{secret}": {
            "put": {
                "description": "approve for single gift card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gift Card"
                ],
                "summary": "approve gift card",
                "operationId": "approve-gift-card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uun",
                        "name": "uun",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "gift card secret",
                        "name": "secret",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GiftCardStatusDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
        "/v1/gift-card/approve-gift-cards": {
            "post": {
                "description": "bulk approve for gift cards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gift Card"
                ],
                "summary": "bulk approve gift cards",
                "operationId": "approve-gift-cards",
                "parameters": [
                    {
                        "description": "bulk approve dto",
                        "name": "approveGiftCardsDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ApproveGiftCardsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GiftCardStatusListDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
        "/v1/gift-card/create-many": {
            "post": {
                "description": "bulk insert for different gift cards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gift Card"
                ],
                "summary": "bulk insert gift cards",
                "operationId": "create-many",
                "parameters": [
                    {
                        "description": "bulk insert gift cards list",
                        "name": "createGiftCards",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkCreateGiftCardsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GiftCardsListDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
        "/v1/gift-card/create-same-many": {
            "post": {
                "description": "bulk insert for the same gift cards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gift Card"
                ],
                "summary": "bulk insert gift cards",
                "operationId": "create-same-many",
                "parameters": [
                    {
                        "description": "bulk insert for the same gift cards dto",
                        "name": "createGiftCards",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.BulkCreateSameGiftCardsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GiftCardsListDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
        "/v1/gift-card/find-by-public-key/{key}": {
            "get": {
                "description": "find a gift card from the db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gift Card"
                ],
                "summary": "Get gift card details by public key",
                "operationId": "find-by-public-key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gift Card public key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GiftCardStatusDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
        "/v1/gift-card/find/{id}": {
            "get": {
                "description": "find a gift card from the db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gift Card"
                ],
                "summary": "Get an gift card by id",
                "operationId": "get-gift-card-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gift Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GiftCardDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
...
        "/v1/gift-card/page/{size}/{number}": {
            "get": {
                "description": "get list of gift cards in paging object",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gift Card"
                ],
                "summary": "gift cards paging",
                "operationId": "find-page",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "size",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "number",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "campaign id",
                        "name": "campaignId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search in public key",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "is valid gift card",
                        "name": "isValid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "expire date from",
                        "name": "expireDateFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "expire date to",
                        "name": "expireDateTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GiftCardsPageDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
        "/v1/gift-card/user-gift-cards/{uun}": {
            "get": {
                "description": "get list of user's gift cards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gift Card"
                ],
                "summary": "user gift cards",
                "operationId": "find-by-uun",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uun",
                        "name": "uun",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GiftCardsListDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/indraframework.IndraException"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CampaignPageDTO": {
            "type": "object",
            "properties": {
                "campaigns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CampaignDTO"
                    }
                },
                "error": {
                    "type": "object",
                    "$ref": "#/definitions/indraframework.IndraException"
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateGiftCardDto": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "expire_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "indraframework.IndraException": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "severity": {
                    "type": "integer"
                },
                "technicalMessage": {
                    "type": "string"
                }
            }
        }
    },
...
}